tests/conform: Fix actor-offscreen-redirect
The actor-offscreen-redirect didn't initialize its state properly, so
it could potentially end up with the "was_painted" state being TRUE
from the start, effectively skipping the whole test.

Fixing so that the test even run resulted in the test getting stuck in
a dead lock due to the verification that a frame was drawn was done
from a paint callback. A paint callback had the mutex held, so when the
test case tried to run the main loop, the next paint callback caller
path taken would try to re-lock the same mutex, thus dead lock.

[Changes originally from the fork of clutter in mutter 3.21.4,
commit 9c0fa583.]

Fixes: https://gitlab.gnome.org/GNOME/clutter/issues/8
Bug-Debian: https://bugs.debian.org/931921
Signed-off-by: Simon McVittie <smcv@debian.org>