tests: Set locale in test-script.c
So that if the translatable strings in test-script.json are ever
translated in the .po files, they will appear translated in the test
program.

Note that for them to be translated in the .po files, they must first be
extracted by gettext, which first needs to grow support for
ClutterScript files.

https://bugzilla.gnome.org/show_bug.cgi?id=770288