meson: Add rules for generating the documentation
Both the API reference and the Cookbook.

This requires a little bit of surgery on the Autotools build as well,
namely that we need to stop generating the main gtk-doc XML file from
a template at configuration time; we should, instead, generate and
include a gtkdocentities.ent file both for the Meson build (where we
generate the file ourselves) and for the Autotools build (where
gtk-doc.make does it for us).